requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
duckdb>=0.9.0
duckdb-engine>=0.17.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# AI/ML
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0 